backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.2
llama-index==0.12.0
llama-index-embeddings-azure-openai==0.3
llama-index-llms-azure-openai==0.3
llama-index-vector-stores-qdrant==0.4
qdrant-client==1.12.2
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pyodbc==5.0.1
python-multipart==0.0.9 
gunicorn == 21.2.0
sqlalchemy==2.0.23